refactor: decompose oversized files to meet 400-line limit
Extract GraphQL unresolvable types to run-unresolvable.ts (Pothos
objectField pattern avoids modifying already-oversized run.ts).
Extract MCP type definitions to formatters-types.ts (-80 lines).
Extract RunDetail banner functions and mutation handlers to
RunDetailBanners.tsx and useRunDetailHandlers.ts (-260 lines).
