+++ cloud/apps/api/src/services/mcp/formatters.ts
@@ -9,88 +9,8 @@ import type { Definition, Run, Transcript, AnalysisResult } from '@prisma/client
 import type { JsonValue } from '@prisma/client/runtime/library';
 import { resolveTranscriptDecisionModel } from '../../graphql/queries/domain/decision-model.js';
 
-/**
- * Definition list item shape for MCP responses
- */
-export type DefinitionListItem = {
-  id: string;
-  name: string;
-  versionLabel: string | null;
-  parentId: string | null;
-  createdAt: string;
-  childCount?: number;
-};
-
-/**
- * Run list item shape for MCP responses
- */
-export type RunListItem = {
-  id: string;
-  status: 'pending' | 'running' | 'completed' | 'failed';
-  models: string[];
-  scenarioCount: number;
-  samplePercentage: number | null;
-  createdAt: string;
-};
-
-/**
- * Run summary shape for MCP responses
- */
-export type RunSummary = {
-  runId: string;
-  status: string;
-  basicStats: {
-    modelCount: number;
-    transcriptCount: number;
-    perModel: Record<string, {
-      sampleSize: number;
-      meanScore: number;
-      stdDev: number;
-    }>;
-  };
-  modelAgreement: {
-    averageCorrelation: number;
-    outlierModels: string[];
-  };
-  mostContestedScenarios: Array<{
-    scenarioId: string;
-    variance: number;
-  }>;
-  insights?: string[];
-  llmSummary?: string;
-  analysisStatus: 'pending' | 'completed' | 'failed';
-  unresolvable: { total: number; byModel: { modelId: string; count: number }[] };
-};
-
-/**
- * Transcript summary shape for MCP responses
- */
-export type TranscriptSummary = {
-  runId: string;
-  scenarioId: string;
-  model: string;
-  turnCount: number;
-  wordCount: number;
-  decision: {
-    direction: string;
-    strength: string;
-    favoredValueKey: string | null;
-  } | null;
-  keyReasoning: string[];
-};
-
-/**
- * Dimension analysis shape for MCP responses
- */
-export type DimensionAnalysis = {
-  rankedDimensions: Array<{
-    name: string;
-    effectSize: number;
-    rank: number;
-  }>;
-  correlations: Record<string, number>;
-  mostDivisive: string[];
-};
+export type { DefinitionListItem, RunListItem, RunSummary, TranscriptSummary, DimensionAnalysis } from './formatters-types.js';
+import type { DefinitionListItem, RunListItem, RunSummary, TranscriptSummary, DimensionAnalysis } from './formatters-types.js';
 
 /**
  * Helper to safely extract array from JSON value
